third_party/WebKit/LayoutTests/transitions/blendmode-transitions-expected.txt

CONSOLE MESSAGE: line 141: Uncaught TypeError: Cannot read property 'cssValueType' of null
Test to make sure that the blendmode property is properly propagated to shared styles
BOX
